Review_physics_2
- refactor how particles were emitted.
- add object pooler to have a better control of memory management;
- add new random helper;
- fews optmizations;

+++ Framework/Graphics/Src/FontRenderer.cpp
@@ -26,14 +26,14 @@ void FontRenderer::DrawScreenText(std::wstring_view str, const Math::Vector2& po
 void Omega::Graphics::FontRenderer::DrawScreenText(TextCommand&& textCommandArgs)
 {
 	auto& [text, color, fontSize, x, y] = textCommandArgs;
-	const size_t length = text.size();
+	const int length = static_cast<int>(text.size());
 	std::string strCopy;
 	std::transform(text.begin(), text.end(), std::back_inserter(strCopy), [](wchar_t c) 
 		{
 			return static_cast<char>(c);
 		});
 	const int convertResult = MultiByteToWideChar(CP_UTF8, 0, strCopy.c_str(), length, nullptr, 0);
-	std::wstring wideString(convertResult, 0);
+	std::wstring wideString(convertResult, 0u);
 	MultiByteToWideChar(CP_UTF8, 0ul, strCopy.c_str(), length, &wideString.at(0), convertResult);
 	TextCommands.emplace_back(std::move(wideString), fontSize, x, y, color);
 }
